project6/requirements.txt
streamlit 
langgraph
langchain
python-dotenv
langchain-ollama